refactor(core): introduce the core module
