+++ mwc/bots/twitter.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 import tweepy
-from mwc.db.models import Movie
+from mwc.core.db.models import Movie
 
 
 class TwitterClient:
